+++ v2/pkg/protocols/common/expressions/variables.go
@@ -4,6 +4,9 @@ import (
 	"errors"
 	"regexp"
 	"strings"
+
+	"github.com/Knetic/govaluate"
+	"github.com/projectdiscovery/nuclei/v2/pkg/operators/common/dsl"
 )
 
 var (
@@ -110,4 +113,13 @@ func ContainsVariablesWithIgnoreList(skipNames map[string]interface{}, items ...
 	}
 
 	return nil
+}
+
+func hasLiteralsOnly(data string) bool {
+	expr, err := govaluate.NewEvaluableExpressionWithFunctions(data, dsl.HelperFunctions)
+	if err == nil && expr != nil {
+		_, err = expr.Evaluate(nil)
+		return err == nil
+	}
+	return true
 }
